move Block to core
Additionally, move LaxBlock to testing
