+++ eras/alonzo/test-suite/test/Test/Cardano/Ledger/Alonzo/Trials.hs
@@ -58,6 +58,7 @@ import Cardano.Ledger.Alonzo.Scripts (ppScript)
 import Cardano.Ledger.Alonzo.Tx (ValidatedTx)
 import Cardano.Ledger.Alonzo.TxBody ()
 import Cardano.Ledger.BaseTypes (Globals)
+import Cardano.Ledger.Block (Block)
 import Cardano.Ledger.Coin (Coin (..))
 import qualified Cardano.Ledger.Core as Core
 import Cardano.Ledger.Era (Era (Crypto))
@@ -65,7 +66,6 @@ import Cardano.Ledger.Hashes (EraIndependentData)
 import Cardano.Ledger.Pretty (PDoc, PrettyA (..))
 import Cardano.Ledger.SafeHash (SafeHash, hashAnnotated)
 import Cardano.Ledger.Shelley (ShelleyEra)
-import Cardano.Ledger.Shelley.BlockChain (Block)
 import Cardano.Ledger.Shelley.Constraints
   ( TransValue,
     UsesAuxiliary,
@@ -88,7 +88,8 @@ import Cardano.Ledger.Shelley.Rules.Delegs (DelegsEnv)
 import Cardano.Ledger.Shelley.Rules.Delpl (DelplEnv, DelplPredicateFailure)
 import Cardano.Ledger.Shelley.Rules.Ledger (LedgerEnv (..))
 import Cardano.Ledger.Shelley.Rules.Utxo (UtxoEnv)
-import Cardano.Ledger.Shelley.TxBody (DCert, TxIn)
+import Cardano.Ledger.Shelley.TxBody (DCert)
+import Cardano.Ledger.TxIn (TxIn)
 import Cardano.Protocol.TPraos.BHeader (BHeader)
 import Cardano.Slotting.Slot (SlotNo (..))
 import Control.Monad.Trans.Reader (runReaderT)
