+++ libs/cardano-ledger-core/cardano-ledger-core.cabal
@@ -45,6 +45,7 @@ library
     Cardano.Ledger.AuxiliaryData
     Cardano.Ledger.BaseTypes
     Cardano.Ledger.BHeaderView
+    Cardano.Ledger.Block
     Cardano.Ledger.Coin
     Cardano.Ledger.Compactible
     Cardano.Ledger.Core
